move filter client from types to extensions (#704)
* move filter client from types to extensions

* update changelog
